insurance-scraper/requirements.txt
selenium>=4.15.0
supabase>=2.0.0
python-dotenv>=1.0.0
webdriver-manager>=4.0.0
requests>=2.31.0
